requirements.txt
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.1.0
asgiref==3.3.4
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.7.1
bleach==3.3.1
certifi==2018.8.24
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.5.0
Deprecated==1.2.12
distlib==0.3.1
Django==3.2.9
django-allauth==0.44.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
entrypoints==0.3
feedparser==6.0.8
filelock==3.0.12
idna==2.7
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mysqlclient==2.1.0
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
oauthlib==3.1.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.17.3
ptyprocess==0.7.0
pycparser==2.20
PyGithub==1.51
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2018.9
pyzmq==22.3.0
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml.clib==0.2.6
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.3.0
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
Unidecode==1.3.2
uritemplate==4.1.1
urllib3==1.25.11
virtualenv==20.4.6
wcwidth==0.2.5
webencodings==0.5.1
Willow==1.3
wrapt==1.12.1
zipp==3.6.0
